+++ src/containers/Map/Layers/ControlLayer/MapFilters/FilterButton/FilterButton.tsx
@@ -9,7 +9,7 @@ interface FilterButtonProps {
 export const FilterButton = ({ onClick }: FilterButtonProps): JSX.Element => {
     return (
         <Button onClick={onClick} className={styles.button}>
-          <img className={styles.image} src={btnBgSrc} alt="Pokaż filtry" />
+            <img className={styles.image} src={btnBgSrc} alt="Pokaż filtry" />
         </Button>
     );
 };