+++ src/containers/Map/Layers/EventLayer/Circle/Circle.ts
@@ -1,11 +1,9 @@
 import * as L from 'leaflet';
 
-const Circle = L.circleMarker([0, 0], {
+export const Circle = L.circleMarker([0, 0], {
   radius: 6, 
   color: '#ff6600', 
   fillColor: 'white', 
   fillOpacity: 1, 
   className: 'mapCircle'
 });
-
-export default Circle;