+++ src/containers/Map/Layers/EventLayer/EventLayer.tsx
@@ -1,7 +1,8 @@
 import { useContext, useEffect } from 'react';
+import { LeafletEvent } from 'leaflet';
 import { MapContext } from 'context/MapContext';
-import createRoutePopup from './RoutePopup/RoutePopup';
-import circle from './Circle/Circle';
+import { createRoutePopup } from './RoutePopup/RoutePopup';
+import { Circle } from './Circle/Circle';
 import './EventLayer.css';
 
 export const EventLayer = (): JSX.Element => {
@@ -13,41 +14,51 @@ export const EventLayer = (): JSX.Element => {
       return;
     }
 
-    map.on('almost:over', (e) => {
-      if (e.layer.options.opacity !== 0){
-        circle.bindTooltip(e.layer.feature.properties.street, {
-          className: 'mouseOverTooltip',
-        });
-        map.addLayer(circle);
-
-        e.layer.setStyle({ color: '#ff6600', dashArray: 6 });
-      }
-    });
-
-    map.on('almost:move', function (e: any) {
-      const pointerPosition = e.latlng;
-      circle.openTooltip();
-
-      circle.setLatLng(pointerPosition);
-    });
-
-    map.on('almost:out', function (e) {
-      circle.unbindTooltip();
-      map.removeLayer(circle);
-      e.layer.setStyle({ color: e.layer.feature.currentColor, dashArray: 0 });
-    });
-
-    map.on('almost:click', function (e) {
-      if (e.layer.options.opacity !== 0){
-        e.layer.unbindPopup();
-        const popup = createRoutePopup('bikelane', e.layer.feature.properties);
-        e.layer.bindPopup(popup, {
-          className: 'popup',
-          minWidth: 200,
-        }).openPopup();
-      }
-    });
+    map.on('almost:over', onMouseOverBikelane);
+    map.on('almost:move', onMouseMoveOverBikelane);
+    map.on('almost:out', onMouseLeaveBikelane);
+    map.on('almost:click', onMouseClickOnBikelane);
   }, []);
+
+  const onMouseOverBikelane = (event: LeafletEvent): void => {
+    if (map !== null && event.layer.options.opacity !== 0){
+      Circle.bindTooltip(event.layer.feature.properties.street, {
+        className: 'mouseOverTooltip',
+      });
+      map.addLayer(Circle);
+
+      // we cant' use event.propagatedFrom until we rewrite almost-over plugin to use it
+      event.layer.setStyle({ color: '#ff6600', dashArray: 6 });
+    }
+  };
+
+  const onMouseMoveOverBikelane = (event: any): void => {
+    const pointerPosition = event.latlng;
+    Circle.openTooltip();
+
+    Circle.setLatLng(pointerPosition);
+  };
+
+  const onMouseLeaveBikelane = (event: LeafletEvent): void => {
+    if (map !== null) {
+      Circle.unbindTooltip();
+      map.removeLayer(Circle);
+      event.layer.setStyle({ color: event.layer.feature.currentColor, dashArray: 0 });
+    }
+  };
+
+  const onMouseClickOnBikelane = (event: LeafletEvent): void => {
+    if (event.layer.options.opacity !== 0){
+      event.layer.unbindPopup();
+
+      const popup = createRoutePopup('bikelane', event.layer.feature.properties);
+
+      event.layer.bindPopup(popup, {
+        className: 'popup',
+        minWidth: 200,
+      }).openPopup();
+    }
+  };
 
   return (
     <></>